third_party/WebKit/LayoutTests/inspector/console/console-css-warnings.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>

<style>
x {}

x {
    #
}

x {
    color: red;
    {}}

x {
    ### {}}

x {
   color: red;
   #########
}

x {
   color: red {};
   color: #;
}

x {
   color: red;
   ###;
   color: @;
}

x {
    color{};
    color{#};
    color{###};
    color{{{}}};
}

x {
    *zoom: @1;  /* IE hack. Ignored */
    color: @;
}

x {
    filter:progid:DXImageTransform.Microsoft.Alpha(opacity=0);  /* IE hack. Ignored */
    filter:alpha(opacity=95);

    background-image:-webkit-linear-gradient(bottom,#2d2927,rgba(49,46,45,.75) 40%,rgba(51,50,49,0) 72%,rgba(54,54,53,0));  /* valid */
    background-image:-webkit-linear-gradient(10);  /* Should be reported. */
    /* Others should be ignored as a vendor-specific extension. */
    background:-moz-linear-gradient(top,rgba(0,0,0,0) 0,rgba(0,0,0,1) 100%);
    background:-webkit-gradient(linear,left top,left bottom,color-stop(0%,rgba(0,0,0,0)),color-stop(100%,rgba(0,0,0,1)));
    background:-webkit-linear-gradient(top,rgba(0,0,0,0) 0,rgba(0,0,0,1) 100%);
    background:-o-linear-gradient(top,rgba(0,0,0,0) 0,rgba(0,0,0,1) 100%);
    background:-ms-linear-gradient(top,rgba(0,0,0,0) 0,rgba(0,0,0,1) 100%);
    background:linear-gradient(to bottom,rgba(0,0,0,0) 0,rgba(0,0,0,1) 100%);
    cursor: hand
}

x {
    /* To be reported. */
    invalidproperty: blue;
    --invalidproperty: blue;
    -unknownvendor-property: blue;

    /* To be ignored. */
    -webkit-invalidproperty: blue;
    scrollbar-arrow-color: fail;
    SCROLLBAR-ARROW-COLOR: fail;
}

x {
    background:         invalid-linear-gradient(top,rgba(0,0,0,0) 0,rgba(0,0,0,1) 100%)    ;
    background:         linear-gradient(top,rgba(0,0,0,0) 0,rgba(0,0,0,1) 100%)    ;
    background: IE-hack ignored \9  ;
    background: IE-hack ignored \0 ;
    background: IE-hack ignored\0/ ;
    background: IE-hack ignored !ie;
    animation: 3s slidein;
    font-size-adjust: none;
    transform: translateZ(0);
    user-select: none;
    -webkit-text-size-adjust: 100%;
    unicode-bidi: isolate;
}

x, {
}

@supports (display: #) { /* No warning expected. */
}

@supports (#) {
}

@supports #;

@supports (display: none) # (display: block) {
}

@xxx;
@xxx {}

@namespace "xxx" %;
@namespace %;
@namespace;

@media %%%;



@-moz-keyframes { }

@page;
@page xxx;

@font-face;
@font-face xxx;

@viewport;
@viewport xxx;

@-webkit-region;
@-webkit-region xxx;

@-webkit-filter;
@-webkit-filter xxx;

@media screen# {}
@media screen, screen# {}
@media screen, screen#, screen {}

@-webkit-keyframes x {
    0x,100% {}
    0%,1x {}
    x
}

div.search input:invalid-pseudo-one {}
div.search input::invalid-pseudo-two {}

div.search input::-webkit-input-placeholder {}
div.search input:-moz-placeholder {}
div.search input:-ms-input-placeholder {}

.valid-vendor-secific-prefixes {
    -ah-x: ?; -Ah-x: ?;
    -atsc-x: ?; 
    -epub-x: ?;
    -hp-x: ?;
    -ibooks-x: ?;
    -khtml-x: ?;
    -ms-x: ?;
    -mso-x: ?;
    -moz-x: ?; -moZ-x: ?;
    -rim-x: ?;
    -ro-x: ?;
    -tc-x: ?;
    -wap-x: ?; -wAp-x: ?;
    -weasy-x: ?;
    -xv-x: ?;
}

.invalid-vendor-specific-prefix {
    -ibookss-x: ?;
    -ibooks-: ?;
    -ibookz-x: ?;
    -m-x: ?;
}
</style>

<script>

function test()
{
    // Ensure script generated errors don't go to the console.
    var style = document.createElement('style');
    style.textContent = "x{y:*}";
    document.head.appendChild(style);

    var messages = WebInspector.consoleModel.messages();
    for (var i = 0; i < messages.length; i++) {
        if (messages[i].source === WebInspector.ConsoleMessage.MessageSource.CSS)
            InspectorTest.addResult(messages[i].message + " (line: " + messages[i].line + ")");
    }

    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p id="p">Tests CSS warnings are properly exposed.</p>
</body>
</html>
