third_party/WebKit/LayoutTests/inspector/console/console-css-warnings-expected.txt

Tests CSS warnings are properly exposed.

Invalid CSS property declaration at: # (line: 10)
Invalid CSS property declaration at: { (line: 15)
Invalid CSS property declaration at: # (line: 18)
Invalid CSS property declaration at: # (line: 22)
Invalid CSS property value: red {} (line: 26)
Invalid CSS property value: # (line: 27)
Invalid CSS property declaration at: # (line: 32)
Invalid CSS property value: @ (line: 33)
Invalid CSS property declaration at: { (line: 37)
Invalid CSS property declaration at: { (line: 38)
Invalid CSS property declaration at: { (line: 39)
Invalid CSS property declaration at: { (line: 40)
Invalid CSS property value: @ (line: 45)
Invalid CSS property value: -webkit-linear-gradient(10) (line: 53)
Invalid CSS property name: invalidproperty (line: 66)
Invalid CSS property declaration at: - (line: 67)
Invalid CSS property name: -unknownvendor-property (line: 68)
Invalid CSS property name: -webkit-invalidproperty (line: 71)
Invalid CSS property value: invalid-linear-gradient(top,rgba(0,0,0,0) 0,rgba(0,0,0,1) 100%) (line: 77)
Invalid CSS selector: x, (line: 91)
Invalid CSS @supports condition: # (line: 97)
Invalid CSS @supports condition: # (line: 100)
Invalid CSS @supports condition: # (line: 102)
Invalid CSS rule at: @xxx (line: 105)
Invalid CSS rule at: @xxx (line: 106)
Invalid CSS media query: %%% (line: 112)
Invalid CSS rule at: ; (line: 118)
Invalid CSS rule at: ; (line: 119)
Invalid CSS rule at: ; (line: 121)
Invalid CSS rule at: xxx (line: 122)
Invalid CSS rule at: ; (line: 124)
Invalid CSS rule at: xxx (line: 125)
Invalid CSS rule at: ; (line: 127)
Invalid CSS rule at: xxx (line: 128)
Invalid CSS rule at: ; (line: 130)
Invalid CSS rule at: ; (line: 131)
Invalid CSS rule at: ; (line: 133)
Invalid CSS rule at: ; (line: 134)
Invalid CSS media query: screen# (line: 136)
Invalid CSS media query: screen# (line: 137)
Invalid CSS media query: screen# (line: 138)
Invalid CSS keyframe selector: 0x,100% (line: 141)
Invalid CSS keyframe selector: 0%,1x (line: 142)
Invalid CSS keyframe selector: x (line: 143)
Invalid CSS selector pseudoclass: invalid-pseudo-one (line: 146)
Invalid CSS selector pseudoclass: invalid-pseudo-two (line: 147)
Invalid CSS property name: -ibookss-x (line: 172)
Invalid CSS property name: -ibooks- (line: 173)
Invalid CSS property name: -ibookz-x (line: 174)
Invalid CSS property name: -m-x (line: 175)
